requirements.txt
# Core LangChain / LangGraph
langgraph>=0.2.50
langchain>=0.3.10
langchain-openai>=0.2.10
langchain-community>=0.3.10
langchain-chroma>=0.1.4

# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.1.3

# Pydantic / config
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1

# Financial data sources
yfinance>=0.2.50
alpha-vantage>=3.0.0
newsapi-python>=0.2.7
tavily-python>=0.5.0

# Vector DB
chromadb>=0.6.0

# SEC filings ingestion
sec-edgar-downloader>=5.0.4
pypdf>=5.1.0
unstructured[local-inference]>=0.16.0; sys_platform != "win32"
unstructured>=0.16.0; sys_platform == "win32"

# Sentiment analysis (FinBERT)
transformers>=4.47.0
torch>=2.5.0

# HTTP + utilities
httpx>=0.28.0
tenacity>=9.0.0
structlog>=24.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
